requirements.txt
pyautogen>=0.4,<0.5
trendspy>=0.1.6
python-dotenv>=1.0.0
openai>=1.0.0
rich>=13.0.0
